pyproject.toml
[tool.poetry]
name = "test-task"
version = "0.1.0"
description = ""
authors = ["xilotttr@gmail.com"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.1.7"
dotenv = "^0.9.9"
psycopg2-binary = "^2.9.10"
djangorestframework = "^3.16.0"
drf-yasg = "^1.21.10"
django-debug-toolbar = "^5.1.0"
django-filter = "^25.1"
drf-nested-routers = "^0.94.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^4.0.1"
flake8 = "^7.1.1"
isort = "^5.13.2"
pylint = "^3.3.2"
black = "^24.10.0"
pytest-cov = "^6.0.0"
pytest-django = "^4.9.0"
django-stubs = "^5.1.1"
mypy = "^1.13.0"
freezegun = "^1.5.1"
pytest = "^8.3.4"
faker = "^36.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
